# be/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Private Clinic API
  version: 0.17.0
  description: Public contract for the private clinic management system.
servers:
  - url: http://localhost:5000
paths:
...
  /api/v1/admin/patients/reference-data:
    get:
      operationId: getPatientReferenceData
      summary: List branches where the actor may manage patient administrative records
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Manageable branches.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientReferenceResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/patients/search:
    post:
      operationId: searchPatients
      summary: Search up to 50 administrative records in one authorized branch
      description: POST keeps phone and national ID search terms out of URLs.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientSearchRequest' }
      responses:
        '200':
          description: Scoped matches.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientSummaryListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/patients/duplicates:
    post:
      operationId: findPatientDuplicates
      summary: Find likely duplicates within the selected branch before registration
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientDuplicateRequest' }
      responses:
        '200':
          description: Up to ten potential matches.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientSummaryListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/patients:
    post:
      operationId: createPatient
      summary: Register a patient after a scoped duplicate check
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePatientRequest' }
      responses:
        '201':
          description: Patient registered.
          headers:
            ETag: { $ref: '#/components/headers/CatalogEtag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/patients/{patientId}:
    get:
      operationId: getPatient
      summary: Read an administrative record in its registration branch
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/BranchPublicIdQuery'
      responses:
        '200':
          description: Patient administrative details.
          headers:
            ETag: { $ref: '#/components/headers/CatalogEtag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientDetailResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updatePatient
      summary: Update administrative details with optimistic concurrency
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/BranchPublicIdQuery'
        - $ref: '#/components/parameters/CatalogIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientWriteRequest' }
      responses:
        '200':
          description: Patient updated.
          headers:
            ETag: { $ref: '#/components/headers/CatalogEtag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /api/v1/admin/patients/{patientId}/merge-preview:
    post:
      operationId: previewPatientMerge
      summary: Preview a duplicate merge for a global administrator
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientMergePreviewRequest' }
      responses:
        '200':
          description: Source, target and impacted record counts.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientMergePreviewResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/patients/{patientId}/merge:
    post:
      operationId: mergePatients
      summary: Merge a duplicate into a canonical patient while preserving immutable history
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/CatalogIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientMergeRequest' }
      responses:
        '200':
          description: Merge completed and audited.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientMergeResultResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/admin/patients/{patientId}/merge-history:
    get:
      operationId: listPatientMergeHistory
      summary: Read audited merge history for a canonical patient
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Merge events ordered from newest to oldest.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientMergeHistoryResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/patients/{patientId}/emergency-contacts:
    put:
      operationId: replacePatientEmergencyContacts
      summary: Replace active emergency contacts with optimistic concurrency and audit
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/BranchPublicIdQuery'
        - $ref: '#/components/parameters/CatalogIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientEmergencyContactsRequest' }
      responses:
        '200':
          description: Active emergency contacts replaced.
          headers:
            ETag: { $ref: '#/components/headers/CatalogEtag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /api/v1/patients/{patientId}/my-emergency-contacts:
    get:
      operationId: getMyEmergencyContacts
      summary: Read emergency contacts for a verified self-linked patient
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Current emergency contacts and patient version.
          headers:
            ETag: { $ref: '#/components/headers/CatalogEtag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MyEmergencyContactsResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: replaceMyEmergencyContacts
      summary: Replace own emergency contacts with optimistic concurrency
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/CatalogIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientEmergencyContactsRequest' }
      responses:
        '200':
          description: Emergency contacts replaced.
          headers:
            ETag: { $ref: '#/components/headers/CatalogEtag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MyEmergencyContactsResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /api/v1/patients/{patientId}/allergies:
    post:
      operationId: addPatientAllergy
      summary: Record an allergy during an active care relationship
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/BranchPublicIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientAllergyRequest' }
      responses:
        '201':
          description: Allergy recorded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientHealthRecordActionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/patients/{patientId}/allergies/{allergyId}/deactivate:
    post:
      operationId: deactivatePatientAllergy
      summary: Deactivate an allergy with an audited reason
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/BranchPublicIdQuery'
        - name: allergyId
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientHealthReasonRequest' }
      responses:
        '200':
          description: Allergy deactivated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientHealthRecordActionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/patients/{patientId}/conditions:
    post:
      operationId: addPatientCondition
      summary: Record a chronic condition during an active care relationship
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/BranchPublicIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientConditionRequest' }
      responses:
        '201':
          description: Condition recorded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientHealthRecordActionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/patients/{patientId}/conditions/{conditionId}/resolve:
    post:
      operationId: resolvePatientCondition
      summary: Resolve a condition with an audited reason
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/BranchPublicIdQuery'
        - name: conditionId
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientHealthReasonRequest' }
      responses:
        '200':
          description: Condition resolved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientHealthRecordActionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/patients/{patientId}/clinical-summary:
    get:
      operationId: getPatientClinicalSummary
      summary: Read allergies and conditions for an active assigned care relationship
      description: Only an assigned doctor or nurse who created the open encounter can read. This endpoint does not return internal results or prescriptions to patients.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/BranchPublicIdQuery'
      responses:
        '200':
          description: Clinical safety summary; read is audited.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientClinicalSummaryResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/reception/patients:
    get:
      operationId: searchReceptionPatients
      summary: Search active patients for walk-in reception at an authorized branch
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BranchPublicIdQuery'
        - name: query
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 100 }
      responses:
        '200':
          description: Matching active patients.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReceptionPatientListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  parameters:
...
    PatientId:
      name: patientId
      in: path
      required: true
      schema: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
...
    CatalogIfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing the Base64 rowversion of a catalog resource.
      schema:
        type: string
        pattern: '^"[A-Za-z0-9+/]{11}="$'
...
    BranchPublicIdQuery:
      name: branchPublicId
      in: query
      required: true
      schema:
        type: string
        pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$'
...
  headers:
    CatalogEtag:
      description: Strong ETag containing the Base64 SQL rowversion.
      schema:
        type: string
        pattern: '^"[A-Za-z0-9+/]{11}="$'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication data is missing, invalid, expired, reused, or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: The account is unavailable for login.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: The requested resource does not exist in the caller's scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: The operation conflicts with uniqueness, concurrency, or account safety rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    PreconditionRequired:
      description: The If-Match precondition is required for this update.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ServiceUnavailable:
      description: A required registration dependency is unavailable.
      content:
        application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
...
    PatientReferenceResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/PatientReferenceData' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    PatientWriteRequest:
      type: object
      required: [fullName, dateOfBirth, gender]
      properties:
        fullName: { type: string, minLength: 2, maxLength: 200 }
        dateOfBirth: { type: string, format: date }
        gender: { type: string, enum: [MALE, FEMALE, OTHER] }
        nationalId: { type: [string, 'null'], maxLength: 30 }
        healthInsuranceNo: { type: [string, 'null'], maxLength: 30 }
        phone: { type: [string, 'null'], maxLength: 20 }
        email: { type: [string, 'null'], maxLength: 254 }
        addressLine: { type: [string, 'null'], maxLength: 300 }
        province: { type: [string, 'null'], maxLength: 100 }
    PatientSearchRequest:
      type: object
      additionalProperties: false
      required: [branchPublicId]
      properties:
        branchPublicId: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
        query: { type: string, minLength: 2, maxLength: 100 }
        dateOfBirth: { type: string, format: date }
    PatientDuplicateRequest:
      allOf:
        - $ref: '#/components/schemas/PatientWriteRequest'
        - type: object
          required: [branchPublicId]
          properties:
            branchPublicId: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
    CreatePatientRequest:
      allOf:
        - $ref: '#/components/schemas/PatientDuplicateRequest'
        - type: object
          properties:
            duplicateOverride: { type: boolean, default: false }
            duplicateReason: { type: string, minLength: 10, maxLength: 500 }
    PatientMergePreviewRequest:
      type: object
      additionalProperties: false
      required: [targetPatientId]
      properties:
        targetPatientId: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
    PatientMergeRequest:
      type: object
      additionalProperties: false
      required: [targetPatientId, targetRowVersion, reason]
      properties:
        targetPatientId: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
        targetRowVersion: { type: string, pattern: '^[A-Za-z0-9+/]{11}=$' }
        reason: { type: string, minLength: 20, maxLength: 500 }
...
    PatientMergePreviewResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/PatientMergePreview' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
...
    PatientMergeResultResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/PatientMergeResult' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
...
    PatientMergeHistoryResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/PatientMergeHistoryItem' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
...
    PatientEmergencyContactsRequest:
      type: object
      additionalProperties: false
      required: [contacts]
      properties:
        contacts:
          type: array
          maxItems: 5
          items: { $ref: '#/components/schemas/PatientEmergencyContact' }
...
    MyEmergencyContactsResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/MyEmergencyContacts' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    PatientAllergyRequest:
      type: object
      additionalProperties: false
      required: [allergenName, type, severity]
      properties:
        allergenName: { type: string, minLength: 2, maxLength: 200 }
        type: { type: string, enum: [DRUG, FOOD, ENVIRONMENT, OTHER] }
        severity: { type: string, enum: [MILD, MODERATE, SEVERE, UNKNOWN] }
        reaction: { type: [string, 'null'], maxLength: 500 }
        notedAt: { type: [string, 'null'], format: date }
...
    PatientHealthRecordActionResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [publicId]
          properties:
            publicId: { type: string, pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
...
    PatientSummaryListResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/PatientSummary' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    PatientDetailResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/PatientDetail' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
...
    ResponseMeta:
      type: object
      additionalProperties: true
...